requirements.txt
pdfplumber
python-gitlab
python-dotenv
